fix(migrations): reverte edição retroativa de migrations + migration corretiva NUMERIC(14,2)
O commit 919d025 (auditoria #4 — dinheiro em Decimal) editou o conteúdo de 17
migrations já aplicadas (004…069), trocando DOUBLE PRECISION por NUMERIC(14,2)
diretamente nos arquivos históricos. Isso quebra o checksum do sqlx em
qualquer ambiente que já tenha rodado essas migrations (VersionMismatch),
inclusive produção quando for para lá.

- Reverte os 17 arquivos ao conteúdo original (DOUBLE PRECISION), restaurando
  os checksums esperados.
- Nova migration 075_money_columns_to_numeric.sql faz a conversão real via
  ALTER COLUMN ... TYPE NUMERIC(14,2) USING ..., idempotente tanto em banco
  novo quanto em banco com as colunas antigas.

Validado: banco local resetado e migrado do zero; information_schema confirma
as 25 colunas monetárias como numeric(14,2); clippy --workspace sem warnings.

+++ server/migrations/037_product_discount.sql
@@ -5,5 +5,5 @@
 --   demais kinds).
 ALTER TABLE products
     ADD COLUMN IF NOT EXISTS discount_kind     TEXT,
-    ADD COLUMN IF NOT EXISTS discount_value NUMERIC(14, 2),
+    ADD COLUMN IF NOT EXISTS discount_value    DOUBLE PRECISION,
     ADD COLUMN IF NOT EXISTS discount_min_qty  DOUBLE PRECISION;
